+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 # [unreleased]
 
+# [v0.20.0] 2023-11-02
+
 ## Added
 
 - Added `__repr__` for `AckPdu` class.
+- Added `__repr__` for `KeepAlivePdu` class.
+- Added `__repr__` for `PromptPdu` class.
 - Added a `finished_params` property for the `FinishedPdu` class.
 - Added a `transmission_mode` property for the `AbstractPduBase` class.
 - Renamed `trans_mode` setter and getter properties to `transmission_mode`.